+++ common/src/main/java/com/evandev/reliable_remover/mixin/minecraft/ItemStackMixin.java
@@ -1,13 +1,19 @@
 package com.evandev.reliable_remover.mixin.minecraft;
 
 import com.evandev.reliable_remover.config.RuleManager;
+import net.minecraft.network.chat.Component;
 import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.context.UseOnContext;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+import java.util.List;
 
 @Mixin(ItemStack.class)
 public class ItemStackMixin {
@@ -16,6 +22,13 @@ public class ItemStackMixin {
     private void reliable_remover$blockUseOn(UseOnContext context, CallbackInfoReturnable<InteractionResult> cir) {
         if (RuleManager.isInteractionBlocked((ItemStack) (Object) this)) {
             cir.setReturnValue(InteractionResult.FAIL);
+        }
+    }
+
+    @Inject(method = "getTooltipLines", at = @At("RETURN"))
+    private void reliable_remover$addTooltip(Item.TooltipContext tooltipContext, Player player, TooltipFlag tooltipFlag, CallbackInfoReturnable<List<Component>> cir) {
+        if (RuleManager.isHidden((ItemStack) (Object) this)) {
+            cir.getReturnValue().add(Component.literal("§4Unobtainable"));
         }
     }
 }