+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-* Improved compatibility with EMI when used standalone.
-* Added support for using EMI with TooManyRecipeViewers. 
-* Backend cleanups.
+* Add tooltip to disabled items.