+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-* Add tooltip to disabled items.
+* Removes disabled items from loot tables.
+* Fixed crash when editing config without Cloth Config.